Terrible currency name dictionary solution
This took me like 1.5 hours on its own. I found it icnredibly difficult and confusing to have a dictionary of currency names, using the values of an enum as keys. I tried several different versions of the enum, the names etc. I tried a Map(), which should be the most direct and explicit syntax for this. All of them had weird typing issues, looping issues and/or ridiculous syntax at the consuming end. I gave up because it was making me despressed and angry. This current version isn't tight enough, but it works and lets me move on.