Apply rate conversion to selected currencies
Holy moly that took ages.

I wrote my tests first. Then attempted to compose the programme. Some
issues:
- it took me ages to realise screen.getText('foo') is an explicit match,
not an automatic conversion to RegExp, so it gets tested against a whole
.textContent, rather than _across_ text content. Hence my tests were all
guaranteed to fail (and did, even when the math + string composition was
right)
- when JS encounters 1, it uses 1 rather than 1.0000... You have to
trick it
- I wrote my own rightPad before discovering padEnd. You'd think after
all the bruhaha around the leftPad module on NPM a few years back that
it would have sunk in that evergreen browsers have a padEnd/padStart